+++ extensions/media2/src/main/java/com/google/android/exoplayer2/ext/media2/PlayerWrapper.java
@@ -202,6 +202,9 @@ import java.util.List;
   }
 
   public boolean removePlaylistItem(@IntRange(from = 0) int index) {
+    if (player.getMediaItemCount() <= index) {
+      return false;
+    }
     player.removeMediaItem(index);
     return true;
   }
@@ -367,7 +370,9 @@ import java.util.List;
       case Player.STATE_IDLE:
         return SessionPlayer.PLAYER_STATE_IDLE;
       case Player.STATE_ENDED:
-        return SessionPlayer.PLAYER_STATE_PAUSED;
+        return player.getCurrentMediaItem() == null
+            ? SessionPlayer.PLAYER_STATE_IDLE
+            : SessionPlayer.PLAYER_STATE_PAUSED;
       case Player.STATE_BUFFERING:
       case Player.STATE_READY:
         return playWhenReady
@@ -428,7 +433,9 @@ import java.util.List;
         postOrRun(handler, pollBufferRunnable);
         break;
       case Player.STATE_ENDED:
-        listener.onPlaybackEnded();
+        if (player.getCurrentMediaItem() != null) {
+          listener.onPlaybackEnded();
+        }
         player.setPlayWhenReady(false);
         updateBufferingState(/* isBuffering= */ false);
         break;